refactor: update common public functions
