+++ crates/common/src/allocations.rs
@@ -1,45 +1,15 @@
 // Copyright 2023-, Edge & Node, GraphOps, and Semiotic Labs.
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::{client::SubgraphClient, watcher::new_watcher};
+use alloy::{primitives::Address, primitives::TxHash};
+use indexer_query::allocations_query::{self, AllocationsQuery};
+use indexer_types::Allocation;
 use std::{
     collections::HashMap,
-    str::FromStr,
     time::{Duration, SystemTime, UNIX_EPOCH},
 };
-
-use super::Allocation;
-use crate::{prelude::SubgraphClient, watcher::new_watcher};
-use alloy::primitives::TxHash;
-use indexer_query::allocations_query::{self, AllocationsQuery};
-use thegraph_core::{Address, DeploymentId};
 use tokio::sync::watch::Receiver;
-
-impl TryFrom<allocations_query::AllocationFragment> for Allocation {
-    type Error = anyhow::Error;
-
-    fn try_from(
-        value: allocations_query::AllocationsQueryAllocations,
-    ) -> Result<Self, Self::Error> {
-        Ok(Self {
-            id: Address::from_str(&value.id)?,
-            status: super::AllocationStatus::Null,
-            subgraph_deployment: super::SubgraphDeployment {
-                id: DeploymentId::from_str(&value.subgraph_deployment.id)?,
-                denied_at: Some(value.subgraph_deployment.denied_at as u64),
-            },
-            indexer: Address::from_str(&value.indexer.id)?,
-            allocated_tokens: value.allocated_tokens,
-            created_at_epoch: value.created_at_epoch as u64,
-            created_at_block_hash: value.created_at_block_hash.to_string(),
-            closed_at_epoch: value.closed_at_epoch.map(|v| v as u64),
-            closed_at_epoch_start_block_hash: None,
-            previous_epoch_start_block_hash: None,
-            poi: None,
-            query_fee_rebates: None,
-            query_fees_collected: None,
-        })
-    }
-}
 
 /// An always up-to-date list of an indexer's active and recently closed allocations.
 pub async fn indexer_allocations(
@@ -118,7 +88,7 @@ mod test {
         "https://api.thegraph.com/subgraphs/name/graphprotocol/graph-network-arbitrum";
     use std::str::FromStr;
 
-    use crate::{prelude::SubgraphClient, subgraph_client::DeploymentDetails};
+    use crate::client::{DeploymentDetails, SubgraphClient};
 
     use super::*;
