+++ crates/common/src/attestation/mod.rs
@@ -1,7 +1,8 @@
 // Copyright 2023-, Edge & Node, GraphOps, and Semiotic Labs.
 // SPDX-License-Identifier: Apache-2.0
 
-mod client;
 mod monitor;
+pub mod signer;
 
-pub use client::{DeploymentDetails, Query, QueryVariables, SubgraphClient};
+pub use monitor::attestation_signers;
+pub use signer::{derive_key_pair, AttestationSigner};
