+++ crates/common/src/client/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2023-, Edge & Node, GraphOps, and Semiotic Labs.
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod dispute_manager;
-pub mod signer;
-pub mod signers;
+mod monitor;
+mod subgraph_client;
+
+pub use subgraph_client::{DeploymentDetails, SubgraphClient};
