+++ crates/common/src/lib.rs
@@ -1,23 +1,24 @@
 // Copyright 2023-, Edge & Node, GraphOps, and Semiotic Labs.
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod address;
-pub mod allocations;
-pub mod attestations;
+mod allocations;
+pub mod attestation;
+pub mod client;
+mod dispute_manager;
 pub mod escrow_accounts;
-pub mod subgraph_client;
+pub mod wallet;
 pub mod watcher;
 
 #[cfg(test)]
 mod test_vectors;
 
-pub mod prelude {
-    pub use super::allocations::{
-        monitor::indexer_allocations, Allocation, AllocationStatus, SubgraphDeployment,
-    };
-    pub use super::attestations::{
-        dispute_manager::dispute_manager, signer::AttestationSigner, signers::attestation_signers,
-    };
+pub mod monitors {
+    pub use super::allocations::indexer_allocations;
+    pub use super::attestation::attestation_signers;
+    pub use super::dispute_manager::dispute_manager;
     pub use super::escrow_accounts::escrow_accounts;
-    pub use super::subgraph_client::{DeploymentDetails, Query, QueryVariables, SubgraphClient};
 }
+
+pub use crate::client::{DeploymentDetails, SubgraphClient};
+pub use escrow_accounts::{EscrowAccounts, EscrowAccountsError};
+pub use indexer_types::{Allocation, AllocationStatus, SubgraphDeployment};
