+++ crates/service/src/service/static_subgraph.rs
@@ -4,7 +4,7 @@
 use axum::{body::Bytes, http::HeaderMap, response::IntoResponse, Extension};
 use tracing::warn;
 
-use indexer_common::subgraph_client::SubgraphClient;
+use indexer_common::SubgraphClient;
 
 use super::{indexer_service::IndexerServiceError, IndexerServiceImpl};
