+++ crates/tap-agent/src/agent/sender_allocation.rs
@@ -10,7 +10,7 @@ use alloy::primitives::Address;
 use alloy::{dyn_abi::Eip712Domain, hex::ToHexExt};
 use anyhow::{anyhow, ensure, Result};
 use bigdecimal::{num_bigint::BigInt, ToPrimitive};
-use indexer_common::{escrow_accounts::EscrowAccounts, prelude::SubgraphClient};
+use indexer_common::{escrow_accounts::EscrowAccounts, SubgraphClient};
 use jsonrpsee::{core::client::ClientT, rpc_params};
 use prometheus::{register_counter_vec, register_histogram_vec, CounterVec, HistogramVec};
 use ractor::{Actor, ActorProcessingErr, ActorRef};
@@ -872,10 +872,7 @@ pub mod tests {
         },
     };
     use futures::future::join_all;
-    use indexer_common::{
-        escrow_accounts::EscrowAccounts,
-        subgraph_client::{DeploymentDetails, SubgraphClient},
-    };
+    use indexer_common::{escrow_accounts::EscrowAccounts, DeploymentDetails, SubgraphClient};
     use jsonrpsee::http_client::HttpClientBuilder;
     use ractor::{
         call, cast, concurrency::JoinHandle, Actor, ActorProcessingErr, ActorRef, ActorStatus,
