+++ crates/types/src/lib.rs
@@ -1,11 +1,12 @@
 // Copyright 2023-, Edge & Node, GraphOps, and Semiotic Labs.
 // SPDX-License-Identifier: Apache-2.0
 
+use std::str::FromStr;
+
 use alloy::primitives::U256;
+use indexer_query::allocations_query;
 use serde::{Deserialize, Deserializer};
 use thegraph_core::{Address, DeploymentId};
-
-pub mod monitor;
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Allocation {
@@ -73,6 +74,33 @@ impl<'d> Deserialize<'d> for Allocation {
             created_at_epoch: outer.createdAtEpoch,
             created_at_block_hash: outer.createdAtBlockHash,
             closed_at_epoch: outer.closedAtEpoch,
+            closed_at_epoch_start_block_hash: None,
+            previous_epoch_start_block_hash: None,
+            poi: None,
+            query_fee_rebates: None,
+            query_fees_collected: None,
+        })
+    }
+}
+
+impl TryFrom<allocations_query::AllocationFragment> for Allocation {
+    type Error = anyhow::Error;
+
+    fn try_from(
+        value: allocations_query::AllocationsQueryAllocations,
+    ) -> Result<Self, Self::Error> {
+        Ok(Self {
+            id: Address::from_str(&value.id)?,
+            status: AllocationStatus::Null,
+            subgraph_deployment: SubgraphDeployment {
+                id: DeploymentId::from_str(&value.subgraph_deployment.id)?,
+                denied_at: Some(value.subgraph_deployment.denied_at as u64),
+            },
+            indexer: Address::from_str(&value.indexer.id)?,
+            allocated_tokens: value.allocated_tokens,
+            created_at_epoch: value.created_at_epoch as u64,
+            created_at_block_hash: value.created_at_block_hash.to_string(),
+            closed_at_epoch: value.closed_at_epoch.map(|v| v as u64),
             closed_at_epoch_start_block_hash: None,
             previous_epoch_start_block_hash: None,
             poi: None,
